video_intelligence_service_sdk/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Configuration
GEMINI_API_KEY = os.getenv('GOOGLE_API_KEY_1')
WHISPER_MODEL_SIZE = "tiny"  # tiny, base, small, medium, large

# File paths
UPLOAD_FOLDER = "uploads"
OUTPUT_FOLDER = "outputs"
os.makedirs(UPLOAD_FOLDER, exist_ok=True)
os.makedirs(OUTPUT_FOLDER, exist_ok=True)