video_intelligence_service_sdk/requirements.txt
faster-whisper
ffmpeg-python
pandas
google-generativeai
python-dotenv